QuickChat/requirements.txt
Django==3.0.7
djangorestframework==3.12.1
django-cors-headers==3.5.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
